V5.0.0 adding Build script and Fixing Course implementation
Repo restructure and build script added.  Also changes to the CRS switch

This is a significant restructure of the repo to remove the versioning from the file and  directories.  A build script has been added.
As for the profile, bumping it to V5 in readiness for the CRS change from encoder to toggle which needs to be implemented in Helios

Various changes

Update README.md
Update Build.cmd
Bug in the final rename step.  Now corrected

Add Version Number Display to Profile Start

Build Script Initial Clean-up